+++ zarrs_opendal/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Move integration tests into `tests/`
+
 ## [0.7.0] - 2025-04-06
 
 ### Changed